Refactor Field interface, enable field attributes
